Fix unknown crash bug
writing on a plain .html file crashes for unknown reason, I fixed it with installing the original nvchad distribution and merge the current configs I have. It added a two files, the toml and lazy-lock.json. Problably this is only for the proper installation since everytime I clone this repo to use my configuration, this bug occurs. So Having it some kind of proper installation might fix the problem.